requirements.txt
Django==4.0
djangorestframework==3.12.4
psycopg2-binary==2.9.1
pytz
